Move more modules into sycamore-core crate
