+++ packages/sycamore-core/src/lib.rs
@@ -1,5 +1,15 @@
-//! Core functionality for the Sycamore web framework.
+//! Core functionality for the Sycamore UI framework.
+//!
+//! This crate should not be used directly. Instead, use the `sycamore` crate which re-exports this
+//! crate.
+//!
+//! # Feature Flags
+//!
+//! - `hydrate` - Enables the hydration API.
 
+pub mod component;
 pub mod generic_node;
+pub mod hydrate;
+pub mod noderef;
 pub mod render;
 pub mod view;
